Improve UnitControl component
This update adds rich features and interactions to the UnitControl component.
Originally, this work was in support of adding padding controls to the Cover block.
This update abstracts just the UnitControls from that effort.

UnitControl and NumberControl are now powered by a new lower level primitive: InputControl.

InputControl handles the majority of the interaction and event logic. Components using it, like NumberControl, are able to tap into the state flow and add features like "drag to update".